♻️ Move formatShortcut to utils and use centralized shortcut constant
