+++ src/shortcut.ts
@@ -1,16 +1,7 @@
 import { invoke } from '@tauri-apps/api/core'
 import { WebviewWindow } from '@tauri-apps/api/webviewWindow'
 import { register, unregister, unregisterAll } from '@tauri-apps/plugin-global-shortcut'
-import { get } from './store'
-
-export const DEFAULT_SHORTCUT = 'CommandOrControl+Shift+X'
-
-export function formatShortcut(shortcut: string, mac: boolean): string {
-  return shortcut
-    .replace('CommandOrControl', mac ? '⌘' : 'Ctrl')
-    .replace('Shift', mac ? '⇧' : 'Shift')
-    .replace('Alt', mac ? '⌥' : 'Alt')
-}
+import { DEFAULT_GLOBAL_SHORTCUT, get } from './store'
 
 async function handleShortcut() {
   const selectedText = await invoke('get_selected_text')
@@ -35,15 +26,18 @@ export async function setupGlobalShortcut(shortcut?: string): Promise<string> {
     console.error('Failed to unregisterAll', e)
     // Fallback: explicitly unregister known shortcuts
     const storedShortcut = await get('global_shortcut')
-    for (const s of [DEFAULT_SHORTCUT, storedShortcut]) {
+    for (const s of [DEFAULT_GLOBAL_SHORTCUT, storedShortcut]) {
       if (s) {
-        try { await unregister(s) } catch (_) {}
+        try {
+          await unregister(s)
+        }
+        catch {}
       }
     }
   }
 
   // 2. Determine shortcut to register
-  const shortcutToRegister = shortcut || (await get('global_shortcut')) || DEFAULT_SHORTCUT
+  const shortcutToRegister = shortcut || (await get('global_shortcut')) || DEFAULT_GLOBAL_SHORTCUT
 
   try {
     await register(shortcutToRegister, (event) => {
@@ -55,13 +49,13 @@ export async function setupGlobalShortcut(shortcut?: string): Promise<string> {
   catch (e) {
     console.error(`Failed to register shortcut ${shortcutToRegister}:`, e)
     // 3. Fallback to default if custom fails
-    if (shortcutToRegister !== DEFAULT_SHORTCUT) {
+    if (shortcutToRegister !== DEFAULT_GLOBAL_SHORTCUT) {
       try {
-        await register(DEFAULT_SHORTCUT, (event) => {
+        await register(DEFAULT_GLOBAL_SHORTCUT, (event) => {
           if (event.state === 'Released')
             handleShortcut()
         })
-        return DEFAULT_SHORTCUT
+        return DEFAULT_GLOBAL_SHORTCUT
       }
       catch (fallbackError) {
         console.error('Failed to register fallback shortcut:', fallbackError)
