+++ src/utils.ts
@@ -27,3 +27,10 @@ export function noop() {}
 export function isUndefined(val: any): val is undefined {
   return val === undefined
 }
+
+export function formatShortcut(shortcut: string, mac: boolean): string {
+  return shortcut
+    .replace('CommandOrControl', mac ? '⌘' : 'Ctrl')
+    .replace('Shift', mac ? '⇧' : 'Shift')
+    .replace('Alt', mac ? '⌥' : 'Alt')
+}
